survey-operator/pyproject.toml
[tool.poetry]
name = "survey-operator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kubernetes = "^30.1.0"
kopf = "^1.37.2"
pyyaml = "^6.0.2"
jinja2 = "^3.1.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
